+++ src/boot/ksdk1.1.0/devBME680.h
@@ -52,4 +52,4 @@ WarpStatus	readSensorSignalBME680(WarpTypeMask signal,
 					WarpSignalReliability reliability,
 					WarpSignalNoise noise);
 void		printSensorDataBME680(bool hexModeFlag);
-void            updateSensorDataBME680(int* current_temp, int* current_hum, int* current_gas_res, uint16_t menuI2cPullupValue);
+void            updateSensorDataBME680(uint8_t* current_temp, uint8_t* current_hum, uint16_t* current_gas_res, uint16_t menuI2cPullupValue);
